+++ migrations/drop-action-schema.js
@@ -1,2 +1,0 @@
-// Drop the actionSchema table that is no longer used
-db.actionSchema.drop();