+++ migrations/drop-old-config.js
@@ -1,2 +1,0 @@
-// Remove the old flow configuration table that is no longer used
-db.deltafiConfig.drop();