+++ migrations/migrate-policy-ids.js
@@ -1,1 +1,0 @@
-db.deletePolicy.updateMany({name: null}, [{$set: {"name": "$_id"}}])