+++ migrations/migrate-requeue-count.js
@@ -1,1 +1,0 @@
-db.deltaFile.updateMany({requeueCount: null}, {$set: {"requeueCount": NumberInt("0")}})