+++ migrations/migrate-rule-ids.js
@@ -1,1 +1,0 @@
-db.flowAssignmentRule.updateMany({name: null}, [{$set: {"name": "$_id"}}])